projects/笔记内容提示--读取/requirements.txt
pyyaml
# pywin32
pyperclip